The editor is moved to a separate component
