+++ client-app/src/pages/home/index.tsx
@@ -1,171 +1,16 @@
 import { useTranslation } from "react-i18next"
 import TestComponent from "./TestComponent"
-import RichTextEditor, {
-  BaseKit,
-  Blockquote,
-  Bold,
-  BulletList,
-  Clear,
-  Code,
-  CodeBlock,
-  Color,
-  ColumnActionButton,
-  Emoji,
-  ExportPdf,
-  ExportWord,
-  FontFamily,
-  FontSize,
-  FormatPainter,
-  Heading,
-  Highlight,
-  History,
-  HorizontalRule,
-  Image,
-  ImportWord,
-  Indent,
-  Italic,
-  Katex,
-  LineHeight,
-  Link,
-  MoreMark,
-  OrderedList,
-  SearchAndReplace,
-  SlashCommand,
-  Strike,
-  Table,
-  TaskList,
-  TextAlign,
-  Underline,
-  Video,
-  TableOfContents,
-  Excalidraw,
-  TextDirection,
-  Mention,
-  Attachment,
-  Mermaid,
-} from "reactjs-tiptap-editor"
 
-// Import CSS
-import "reactjs-tiptap-editor/style.css"
 import { useState } from "react"
-import { convertBase64ToBlob } from "@/lib/utils"
 import { Button } from "@/components/ui/button"
+import { Editor } from "@/components/editor"
 
-const extensions = [
-  BaseKit.configure({
-    multiColumn: true,
-    placeholder: {
-      showOnlyCurrent: true,
-    },
-    characterCount: {
-      limit: 50_000,
-    },
-  }),
-  History,
-  SearchAndReplace,
-  TextDirection,
-  TableOfContents,
-  FormatPainter.configure({ spacer: true }),
-  Clear,
-  FontFamily,
-  Heading.configure({ spacer: true }),
-  FontSize,
-  Bold,
-  Italic,
-  Underline,
-  Strike,
-  MoreMark,
-  Katex,
-  Emoji,
-  Color.configure({ spacer: true }),
-  Highlight,
-  BulletList,
-  OrderedList,
-  TextAlign.configure({ types: ["heading", "paragraph"], spacer: true }),
-  Indent,
-  LineHeight,
-  TaskList.configure({
-    spacer: true,
-    taskItem: {
-      nested: true,
-    },
-  }),
-  Link,
-  Image.configure({
-    upload: (files: File) => {
-      return new Promise((resolve) => {
-        setTimeout(() => {
-          resolve(URL.createObjectURL(files))
-        }, 500)
-      })
-    },
-  }),
-  Video.configure({
-    upload: (files: File) => {
-      return new Promise((resolve) => {
-        setTimeout(() => {
-          resolve(URL.createObjectURL(files))
-        }, 500)
-      })
-    },
-  }),
-  Blockquote.configure({ spacer: true }),
-  SlashCommand,
-  HorizontalRule,
-  Code.configure({
-    toolbar: false,
-  }),
-  CodeBlock.configure({ defaultTheme: "dracula" }),
-  ColumnActionButton,
-  Table,
-  ExportPdf.configure({ spacer: true }),
-  ImportWord.configure({
-    upload: (files: File[]) => {
-      const f = files.map((file) => ({
-        src: URL.createObjectURL(file),
-        alt: file.name,
-      }))
-      return Promise.resolve(f)
-    },
-  }),
-  ExportWord,
-  Excalidraw,
-  Mention,
-  Attachment.configure({
-    upload: (file: any) => {
-      // fake upload return base 64
-      const reader = new FileReader()
-      reader.readAsDataURL(file)
-
-      return new Promise((resolve) => {
-        setTimeout(() => {
-          const blob = convertBase64ToBlob(reader.result as string)
-          resolve(URL.createObjectURL(blob))
-        }, 300)
-      })
-    },
-  }),
-  Mermaid.configure({
-    upload: (file: any) => {
-      // fake upload return base 64
-      const reader = new FileReader()
-      reader.readAsDataURL(file)
-
-      return new Promise((resolve) => {
-        setTimeout(() => {
-          const blob = convertBase64ToBlob(reader.result as string)
-          resolve(URL.createObjectURL(blob))
-        }, 300)
-      })
-    },
-  }),
-]
 
 export default function Home() {
   const { t } = useTranslation("translation")
   const [content, setContent] = useState("")
 
-  const handleChangeContent = (value: any) => {
+  const handleChange = (value: any) => {
     setContent(value)
   }
 
@@ -185,11 +30,9 @@ export default function Home() {
       <section className="w-full py-32 md:py-48 flex flex-col items-center justify-center">
         <h1 className="text-black">{t("title")}</h1>
         <TestComponent />
-        <RichTextEditor
-          output="text"
+        <Editor
           content={content}
-          onChangeContent={handleChangeContent}
-          extensions={extensions}
+          onChange={handleChange}
         />
       <Button onClick={handleSave}>Save</Button>
       </section>
